+++ beast-of-argh/beast-of-argh.d/local/share/cook/bin/configure-grafana.sh
@@ -29,6 +29,13 @@ chown -R grafana:grafana /mnt/applog/grafana
 #  INFLUXDBSOURCE="$INFLUXDBSOURCE:8086"
 #fi
 
+# dashboard sources
+# https://raw.githubusercontent.com/lux4rd0/grafana-loki-syslog-aio/main/config/grafana/dashboards/no_folder/loki_syslog_aio_overview.json
+# https://raw.githubusercontent.com/mr-karan/nomad-monitoring/main/dashboards/allocations.json
+# https://raw.githubusercontent.com/mr-karan/nomad-monitoring/main/dashboards/clients.json
+# https://raw.githubusercontent.com/mr-karan/nomad-monitoring/main/dashboards/server.json
+# Todo: pull these in from source
+
 # if /mnt/grafana is empty, copy in /var/db/grafana
 if [ ! -f /mnt/grafana/grafana.db ]; then
     # if empty we need to copy in the directory structure from install
@@ -77,10 +84,12 @@ if [ ! -f /mnt/grafana/grafana.db ]; then
       /mnt/grafana/provisioning/dashboards/newhomelogs.json
     cp "$TEMPLATEPATH/vault.json.in" \
       /mnt/grafana/provisioning/dashboards/vault.json
-    cp "$TEMPLATEPATH/nomadcluster.json.in" \
-      /mnt/grafana/provisioning/dashboards/nomadcluster.json
-    cp "$TEMPLATEPATH/nomadjobs.json.in" \
-      /mnt/grafana/provisioning/dashboards/nomadjobs.json
+    cp "$TEMPLATEPATH/nomadallocations.json.in" \
+      /mnt/grafana/provisioning/dashboards/nomadallocations.json
+    cp "$TEMPLATEPATH/nomadclient.json.in" \
+      /mnt/grafana/provisioning/dashboards/nomadclient.json
+    cp "$TEMPLATEPATH/nomadservers.json.in" \
+      /mnt/grafana/provisioning/dashboards/nomadservers.json
     cp "$TEMPLATEPATH/consulcluster.json.in" \
       /mnt/grafana/provisioning/dashboards/consulcluster.json
     cp "$TEMPLATEPATH/postgres.json.in" \
@@ -134,10 +143,12 @@ else
       /mnt/grafana/provisioning/dashboards/newhomelogs.json
     cp "$TEMPLATEPATH/vault.json.in" \
       /mnt/grafana/provisioning/dashboards/vault.json
-    cp "$TEMPLATEPATH/nomadcluster.json.in" \
-      /mnt/grafana/provisioning/dashboards/nomadcluster.json
-    cp "$TEMPLATEPATH/nomadjobs.json.in" \
-      /mnt/grafana/provisioning/dashboards/nomadjobs.json
+    cp "$TEMPLATEPATH/nomadallocations.json.in" \
+      /mnt/grafana/provisioning/dashboards/nomadallocations.json
+    cp "$TEMPLATEPATH/nomadclient.json.in" \
+      /mnt/grafana/provisioning/dashboards/nomadclient.json
+    cp "$TEMPLATEPATH/nomadservers.json.in" \
+      /mnt/grafana/provisioning/dashboards/nomadservers.json
     cp "$TEMPLATEPATH/consulcluster.json.in" \
       /mnt/grafana/provisioning/dashboards/consulcluster.json
     cp "$TEMPLATEPATH/postgres.json.in" \